add license
LUXCIUM‡ LICENSE
*NO* PERMISSION GRANTED
PROVIDED "AS IS"
WITHOUT WARRANTY

*NO* PERMISSION ARE GRANTED.
NOT TO PUBLISH, NOT TO DISTRIBUTE,
NOT TO SUBLICENSE, AND/OR NOT TO
SELL COPIES OF THE SOFTWARE.
You may use, only for yourself,
the source code in this file and
from imported files to this one.
You have the right to: distribute to your friends, to your students,
or to you coworker for personal or
academic use at home at school or
at work. You must include this
«FULL NOTICE» in each imported
files to this one that you decide
to use if such «FULL NOTICE»
is missing. If you modify the code
in a substatial way you may decide
to include only the «COPYRIGHT»
and the «SHORT NOTICE» below.

† Copyright © 2020 — LUXCIUM‡
(Benjamin Vincent Kasapoglu)<luxcium⸓neb401.com>

THE SOFTWARE IS PROVIDED "AS IS",

WITHOUT WARRANTY OF ALL KIND,
EXPRESS OR IMPLIED, INCLUDING BUT
NOT LIMITED TO THE WARRANTIES OF
MERCHANTABILITY, FITNESS FOR A
PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
THE AUTHORS OR COPYRIGHT HOLDERS
BE LIABLE FOR ALL CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
ACTION OF CONTRACT, TORT OR
OTHERWISE, ARISING FROM, OUT OF
OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
THE SOFTWARE.

- IN ANY CASES THE COPYRIGHT AND SHORT NOTICE ABOVE MUST BE INCLUDED.

† Scientia es lux principium✨ ™
*/

Signed-off-by: Benjamin Vincent (Luxcium) <luxcium@neb401.com>