ninth day (#9) completed late on day 19th
* feat: tag list ordinal english number names 0-100

Signed-off-by: Benjamin Vincent (Luxcium) <luxcium@neb401.com>

* wakatime tracker

Signed-off-by: Benjamin Vincent (Luxcium) <luxcium@neb401.com>

* catching up: working on day two

Signed-off-by: Benjamin Vincent (Luxcium) <luxcium@neb401.com>

* catching up: working on day three

Signed-off-by: Benjamin Vincent (Luxcium) <luxcium@neb401.com>

* catching up: working on day four

Signed-off-by: Benjamin Vincent (Luxcium) <luxcium@neb401.com>

* Scientia es lux principium✨ -- fifth-day

* Scientia es lux principium✨ -- sixth-day

* Scientia es lux principium✨ -- seventh-day

* Scientia es lux principium✨ -- eighth-day

* Scientia es lux principium✨ -- ninth-day

* FEAT: implement document and test of Functor

Signed-off-by: Benjamin Vincent (Luxcium) <luxcium@neb401.com>

* feat: docs and implementation of Functor

Signed-off-by: Benjamin Vincent (Luxcium) <luxcium@neb401.com>

* add license

LUXCIUM‡ LICENSE
*NO* PERMISSION GRANTED
PROVIDED "AS IS"
WITHOUT WARRANTY

*NO* PERMISSION ARE GRANTED.
NOT TO PUBLISH, NOT TO DISTRIBUTE,
NOT TO SUBLICENSE, AND/OR NOT TO
SELL COPIES OF THE SOFTWARE.
You may use, only for yourself,
the source code in this file and
from imported files to this one.
You have the right to: distribute to your friends, to your students,
or to you coworker for personal or
academic use at home at school or
at work. You must include this
«FULL NOTICE» in each imported
files to this one that you decide
to use if such «FULL NOTICE»
is missing. If you modify the code
in a substatial way you may decide
to include only the «COPYRIGHT»
and the «SHORT NOTICE» below.

† Copyright © 2020 — LUXCIUM‡
(Benjamin Vincent Kasapoglu)<luxcium⸓neb401.com>

THE SOFTWARE IS PROVIDED "AS IS",

WITHOUT WARRANTY OF ALL KIND,
EXPRESS OR IMPLIED, INCLUDING BUT
NOT LIMITED TO THE WARRANTIES OF
MERCHANTABILITY, FITNESS FOR A
PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
THE AUTHORS OR COPYRIGHT HOLDERS
BE LIABLE FOR ALL CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
ACTION OF CONTRACT, TORT OR
OTHERWISE, ARISING FROM, OUT OF
OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
THE SOFTWARE.

- IN ANY CASES THE COPYRIGHT AND SHORT NOTICE ABOVE MUST BE INCLUDED.

† Scientia es lux principium✨ ™
*/

Signed-off-by: Benjamin Vincent (Luxcium) <luxcium@neb401.com>

* LUXCIUM LICENSE *NO* PERMISSION GRANTED
PROVIDED "AS IS"
WITHOUT WARRANTY

Signed-off-by: Benjamin Vincent (Luxcium) <luxcium@neb401.com>

* Liscense descriptor

Signed-off-by: Benjamin Vincent (Luxcium) <luxcium@neb401.com>

* Copyright

Signed-off-by: Benjamin Vincent (Luxcium) <luxcium@neb401.com>

* I have work on the functor project many days
for the main part of days 2,3,6,8 and 9

Signed-off-by: Benjamin Vincent (Luxcium) <luxcium@neb401.com>